requirements.txt
discord-py-interactions==5.13.2
tansy==0.9.2
python-dotenv==1.0.1
humanize==4.11.0
orjson==3.10.13; implementation_name == "cpython"
uvloop==0.21.0; platform_system == "Linux" and implementation_name == "cpython"